+++ src/pages/home/quest.js
@@ -1,17 +1,119 @@
 import $ from 'jquery';
 import fetch from 'node-fetch';
+import GamePagesEnum from '../../constants/gamePagesEnum';
 import { ErrorTypesEnum, FarmRPGPlusError } from '../../FarmRPGPlusError';
 import ConsolePlus from '../../modules/consolePlus';
 import { createRow } from '../../modules/rowFactory';
+import SettingsPlus from '../../modules/settingsPlus';
 import StoragePlus from '../../modules/storagePlus';
 import Roman from '../../utils/roman';
 import { createCardList, getListByTitle, parseNameForUrl } from '../../utils/utils';
 
 class QuestPage {
+
+    constructor() {
+        SettingsPlus.registerPage(GamePagesEnum.QUEST, {
+            displayName: 'Quest Page',
+            order: 5,
+        });
+
+        SettingsPlus.registerFeature(
+            GamePagesEnum.QUEST,
+            'addBuddyFarmCard',
+            {
+                title: 'Add Buddy Farm Card?',
+                subtitle: 'Adds a card with links to Buddy Farm quest information.',
+                isEnabled: true,
+                configs: {}
+            }
+        );
+
+        SettingsPlus.registerFeature(
+            GamePagesEnum.QUEST,
+            'addExtraBuddyFarmButtons',
+            {
+                title: 'Add Extra Buddy Farm Buttons?',
+                subtitle: 'Adds buttons for previous and next Buddy Farm quest pages.',
+                isEnabled: true,
+                configs: {}
+            }
+        );
+        
+        SettingsPlus.registerFeature(
+            GamePagesEnum.QUEST,
+            'addLibraryButtonToPhr',
+            {
+                title: 'Add Library Button to PHRs?',
+                subtitle: 'Adds a button to Personal Help Request pages that links to the library.',
+                isEnabled: true,
+                configs: {}
+            }
+        );
+
+    }
+
     static titles = {
         SILVER_REQUESTED: 'Silver Requested',
         ITEMS_REQUESTED: 'Items Requested',
         REWARDS: 'Rewards',
+    };
+
+    createBuddyFarmCardList = (page) => {
+        if (!page?.container) {
+            new FarmRPGPlusError(
+                ErrorTypesEnum.PARAMETER_MISMATCH,
+                this.createBuddyFarmCardList.name,
+            );
+            return null;
+        }
+
+        let frpgpBuddyFarmCardExists = $(page.container).find('#frpgp-buddy-farm-quest-info-card').length > 0;
+        let $bfCardList;
+
+        if (!frpgpBuddyFarmCardExists) {
+            let $bfCard = createCardList({
+                cardId: 'frpgp-buddy-farm-quest-info-card',
+                title: 'Buddy Farm Quest Information',
+                children: [],
+            });
+
+            $bfCard = getListByTitle(
+                page,
+                QuestPage.titles.REWARDS,
+                { returnTitle: true },
+            )
+                .next('.card')
+                .after($bfCard)
+                .next()
+                .next();
+
+            ConsolePlus.debug('Buddy Farm card created:', $bfCard);
+
+            $bfCardList = $bfCard.find('ul');
+
+            ConsolePlus.debug('Buddy Farm card list:', $bfCardList);
+            
+            if ($bfCardList.length === 0) {
+                new FarmRPGPlusError(
+                    ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                    this.createBuddyFarmCardList.name,
+                );
+                return null;
+            }
+
+        } else {
+            $bfCardList = $(page.container).find('#frpgp-buddy-farm-quest-info-card ul');
+
+            if ($bfCardList.length === 0) {
+                new FarmRPGPlusError(
+                    ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                    this.getBuddyFarmCardList.name,
+                );
+                return null;
+            }
+        }
+
+        return $bfCardList;
     };
 
     checkIfNextQuestExists = (questName = '', nextQuestNumber = '') => {
@@ -113,6 +215,55 @@ class QuestPage {
         return;
     };
 
+    addLibraryButtonToPhr = (page) => {
+        if (!page?.container) {
+            new FarmRPGPlusError(
+                ErrorTypesEnum.PAGE_NOT_FOUND,
+                this.addLibraryButtonToPhr.name,
+            );
+            return;
+        }
+
+        if (!this.isPhrQuestPage(page)) {
+            return;
+        }
+
+        if (!SettingsPlus.isEnabled(GamePagesEnum.QUEST, 'addLibraryButtonToPhr')) {
+            ConsolePlus.log('Library button for PHR is disabled in settings.');
+            return;
+        }
+
+        const $peachJuiceBtn = $(page.container).find('.button.btnorange.drinkpj');
+        if ($peachJuiceBtn.length === 0) {
+            ConsolePlus.log('Not a PHR page, skipping library button addition.');
+            return;
+        }
+
+        const $libraryButton = createRow({
+            iconClass: 'fa fa-fw fa-book',
+            title: 'Library - Personal Help Requests',
+            subtitle: 'Open\'s library personal help requests page',
+            rowLink: 'wiki.php?page=Personal+Help+Requests',
+        });
+
+        const $card = createCardList({
+            cardId: 'frpgp-library-info-card',
+            title: 'Library Information',
+            children: [$libraryButton],
+        });
+
+        const $lastTitle = getListByTitle(
+            page,
+            QuestPage.titles.REWARDS,
+            { returnTitle: true },
+        );
+
+        const itExists = $(page.container).find('#frpgp-library-info-card').length > 0;
+        if (!itExists) {
+            $lastTitle.next('.card').after($card);
+        }
+    };
+
     addBuddyFarmCard = (page) => {
         if (!page?.container) {
             new FarmRPGPlusError(
@@ -122,8 +273,16 @@ class QuestPage {
             return;
         }
 
-        const $questTitle = $(page.container).find('.item-title[style=\'font-weight: bold\']');
+        if (this.isPhrQuestPage(page)) {
+            return;
+        }
 
+        if (!SettingsPlus.isEnabled(GamePagesEnum.QUEST, 'addBuddyFarmCard')) {
+            ConsolePlus.log('Buddy Farm card is disabled in settings.');
+            return;
+        }
+
+        const $questTitle = $(page.container).find('.item-title[style=\'font-weight: bold\']');
         if (!$questTitle || $questTitle.length === 0) {
             new FarmRPGPlusError(
                 ErrorTypesEnum.ELEMENT_NOT_FOUND,
@@ -132,7 +291,6 @@ class QuestPage {
             return;
         }
         
-        const pageIsPhr = this.isPhrQuestPage(page);
         const questName = $questTitle.children().length >= 1 // Check if it title its multi-row
             ? $questTitle.html().trim().replace(/<br\s*\/?>/gi, ' ') // Replace <br> tags with spaces
             : $questTitle.text().trim();
@@ -146,46 +304,28 @@ class QuestPage {
             return;
         }
 
-        let overrideData = {
-            title: null,
-            iconClass: null,
-            subtitle: null,
-            rowLink: null,
-        };
-
-        if (pageIsPhr) {
-            overrideData.iconClass = 'fa fa-fw fa-book';
-            overrideData.title = 'Library - Personal Help Requests';
-            overrideData.subtitle = 'Open\'s library personal help requests page';
-            overrideData.rowLink = 'wiki.php?page=Personal+Help+Requests';
-        }
-
         const $bfCurrentQuestRow = createRow({
             rowId: 'frpgp-buddy-farm-current-quest',
-            title: overrideData.title ?? `BF - ${questName}`,
-            iconClass: overrideData.iconClass ?? 'fa fa-fw fa-bullseye',
-            subtitle: overrideData.subtitle ?? 'Open\'s Buddy Farm current quest page',
-            rowLink: overrideData.rowLink ?? `https://buddy.farm/q/${parseNameForUrl(questName)}`,
+            title: `BF - ${questName}`,
+            iconClass: 'fa fa-fw fa-bullseye',
+            subtitle: 'Open\'s Buddy Farm current quest page',
+            rowLink: `https://buddy.farm/q/${parseNameForUrl(questName)}`,
         });
 
-        const $bfCard = createCardList({
-            cardId: 'frpgp-buddy-farm-quest-info-card',
-            title: 'Quest Information',
-            children: [$bfCurrentQuestRow],
-        });
-
-        const $lastTitle = getListByTitle(
-            page,
-            QuestPage.titles.REWARDS,
-            { returnTitle: true },
-        );
-
-        const itExists = $(page.container).find('#frpgp-buddy-farm-quest-info-card');
-        if (itExists.length > 0) {
-            itExists.remove();
+        const itExists = $(page.container).find('#frpgp-buddy-current-quest').length > 0;
+        if (!itExists) {
+            const $frpgpBuddyFarmCard = this.createBuddyFarmCardList(page);
+            ConsolePlus.debug('Buddy Farm card for current quest:', $frpgpBuddyFarmCard);
+            if ($frpgpBuddyFarmCard) {
+                $frpgpBuddyFarmCard.append($bfCurrentQuestRow);
+            } else {
+                new FarmRPGPlusError(
+                    ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                    this.addBuddyFarmCard.name,
+                );
+                return;
+            }
         }
-
-        $lastTitle.next('.card').after($bfCard);
     };
 
     addExtraBuddyFarmButtons = (page) => {
@@ -197,10 +337,13 @@ class QuestPage {
             return;
         }
 
-        const pageIsPhr = this.isPhrQuestPage(page);
-        
-        if (pageIsPhr) {
-            return; // Skip next quest check for PHR pages
+        if (this.isPhrQuestPage(page)) {
+            return;
+        }
+
+        if (!SettingsPlus.isEnabled(GamePagesEnum.QUEST, 'addExtraBuddyFarmButtons')) {
+            ConsolePlus.log('Extra Buddy Farm buttons are disabled in settings.');
+            return;
         }
 
         const $questTitle = $(page.container).find('.item-title[style=\'font-weight: bold\']');
@@ -225,11 +368,13 @@ class QuestPage {
             return;
         }
 
-
         let $previousQuestRow;
         let $nextQuestRow;
 
-        const [previousQuestNumber, nextQuestNumber] = this.getPreviousAndNextQuestNumbers(questName);
+        const [
+            previousQuestNumber,
+            nextQuestNumber
+        ] = this.getPreviousAndNextQuestNumbers(questName);
 
         if (previousQuestNumber) {
             const previousQuestName = questName.replace(
@@ -247,10 +392,19 @@ class QuestPage {
 
             const itExists = $(page.container).find('#frpgp-buddy-farm-previous-quest').length > 0;
             if (!itExists) {
-                $(page.container).find('#frpgp-buddy-farm-current-quest').parent().before($previousQuestRow);
+
+                const $frpgpBuddyFarmCard = this.createBuddyFarmCardList(page);
+                if ($frpgpBuddyFarmCard) {
+                    $frpgpBuddyFarmCard.prepend($previousQuestRow);
+                } else {
+                    new FarmRPGPlusError(
+                        ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                        this.addExtraBuddyFarmButtons.name,
+                    );
+                    return;
+                }
             }
         }
-
 
         this.checkIfNextQuestExists(questName, nextQuestNumber).then(
             (exists) => {
@@ -271,9 +425,19 @@ class QuestPage {
                     rowLink: `https://buddy.farm/q/${parseNameForUrl(nextQuestName)}`,
                 });
 
-                const itExists = $(page.container).find('#frpgp-buddy-farm-next-quest');
-                if (!itExists.length) {
-                    $(page.container).find('#frpgp-buddy-farm-current-quest').parent().after($nextQuestRow);
+                const itExists = $(page.container).find('#frpgp-buddy-farm-next-quest').length > 0;
+                if (!itExists) {
+                    const $frpgpBuddyFarmCard = this.createBuddyFarmCardList(page);
+                    ConsolePlus.debug('Buddy Farm card for next quest:', $frpgpBuddyFarmCard);
+                    if ($frpgpBuddyFarmCard) {
+                        $frpgpBuddyFarmCard.append($nextQuestRow);
+                    } else {
+                        new FarmRPGPlusError(
+                            ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                            this.addExtraBuddyFarmButtons.name,
+                        );
+                        return;
+                    }
                 }
 
                 return true;
@@ -291,6 +455,7 @@ class QuestPage {
         }
 
         ConsolePlus.log('Quest page initialized:', page);
+        this.addLibraryButtonToPhr(page);
         this.addBuddyFarmCard(page);
         this.addExtraBuddyFarmButtons(page);
     };