+++ webpack.config.cjs
@@ -13,6 +13,7 @@ module.exports = {
     plugins: [
         new webpack.DefinePlugin({
             'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'production'),
+            'process.env.VERSION': JSON.stringify(pkg.version),
         }),
 
         new UserscriptPlugin({